extracting payees fragment
