+++ src/com/money/manager/ex/MainActivity.java
@@ -56,7 +56,6 @@ import android.widget.Toast;
 import com.afollestad.materialdialogs.AlertDialogWrapper;
 import com.afollestad.materialdialogs.MaterialDialog;
 import com.money.manager.ex.CategorySubCategoryExpandableListActivity.CategorySubCategoryExpandableLoaderListFragment;
-import com.money.manager.ex.PayeeActivity.PayeeLoaderListFragment;
 import com.money.manager.ex.about.AboutActivity;
 import com.money.manager.ex.adapter.DrawerMenuItem;
 import com.money.manager.ex.adapter.DrawerMenuItemAdapter;
@@ -71,6 +70,7 @@ import com.money.manager.ex.fragment.AccountLoaderListFragment;
 import com.money.manager.ex.fragment.BaseFragmentActivity;
 import com.money.manager.ex.fragment.DashboardFragment;
 import com.money.manager.ex.fragment.HomeFragment;
+import com.money.manager.ex.fragment.PayeeLoaderListFragment;
 import com.money.manager.ex.notifications.RepeatingTransactionNotifications;
 import com.money.manager.ex.preferences.PreferencesConstant;
 import com.money.manager.ex.recurring.transactions.RepeatingTransactionListFragment;
